+++ astropy/nddata/__init__.py
@@ -20,6 +20,7 @@ from .mixins.ndslicing import *
 from .mixins.ndio import *
 
 from .compat import *
+from .utils import *
 
 from .. import config as _config